[ADD] filter for children values after mapping
The filter was previously applied before executing the mapping for the
item values, via the `skip_item` hook.

In the cases that the filter depends of the results of the mapping the
additional hook `skip_add_item` is useful.

It allows to skip the action adding of the item to the result list.

Also the unit tests for both filters have been added.