Obtain a session with a pre-request
This fix the problem explained in the previous commit:
When we run a job, we don't wait for the answer. So when we don't have
any cookie yet, we send a GET on a very fast controller route which will
return a cookie, used for all the subsequents calls to runjob.

This effectively prevent to create one session file per job.

It still needs intensive testing.