runner: Increase login timeout to 60 s
On staging we have the following problem for the first Connector job of
each day:

1. The main server enqueues a job.
2. The runner on the Connector server sends a log-in request to the
   worker processes.
3. One of the worker processes receives it, loads its modules and does
   other initialization. This is because it hadn't received any requests
   before. And it takes a long time, perhaps because the first Connector
   job of the day arrives long after the the Connector server is
   started, so its virtual memory is paged out.
4. The runner times out and restarts itself.
5. The job doesn't get processed, because the runner doesn't
   automatically restart jobs.

Increase the waiting time in order to avoid timeouts due to slow worker
initialization. This should prevent steps 4 and 5 and therefore fix our
problem of jobs getting stuck. I've seen workers taking up to 40 s to
initialize, so 60 s should be just right.

Issue #6080